+++ xcube_hub/core/cubegens.py
@@ -12,7 +12,7 @@ from urllib3.exceptions import MaxRetryError
 from xcube_hub import api, poller
 from xcube_hub.api import get_json_request_value
 from xcube_hub.auth import Auth
-from xcube_hub.core import callbacks, costs
+from xcube_hub.core import callbacks, costs, punits
 from xcube_hub.core import user_namespaces
 from xcube_hub.keyvaluedatabase import KeyValueDatabase
 from xcube_hub.typedefs import AnyDict, Error, JsonObject
@@ -109,12 +109,29 @@ def create_cubegen_object(cubegen_id: str, cfg: AnyDict, info_only: bool = False
     return cubegen
 
 
-def create(user_id: str, cfg: AnyDict, token: Optional[str] = None, info_only: bool = False) -> Union[AnyDict, Error]:
+def _raise_for_invalid_punits(user_id: str, email: str, cfg: AnyDict, token: str):
+    limit = os.getenv("XCUBE_HUB_PROCESS_LIMIT", 1000)
+
+    infos = info(user_id=user_id, email=email, body=cfg, token=token)
+    cost_estimation = infos['cost_estimation']
+
+    if cost_estimation['required'] > limit:
+        raise api.ApiError(400, f"Number of required punits ({cost_estimation['required']}) is greater than the absolute limit of {limit}")
+
+    if cost_estimation['required'] > cost_estimation['available']:
+        raise api.ApiError(400, f"Number of required punits ({cost_estimation['required']}) "
+                                f"is greater than the available {cost_estimation['available']}")
+
+
+def create(user_id: str, email: str, cfg: AnyDict, token: Optional[str] = None, info_only: bool = False) -> Union[AnyDict, Error]:
     try:
         if 'input_config' not in cfg and 'input_configs' not in cfg:
             raise api.ApiError(400, "Either 'input_config' or 'input_configs' must be given")
 
         token = token or Auth.instance().token
+
+        if not info_only:
+            _raise_for_invalid_punits(user_id=user_id, email=email, cfg=cfg, token=token)
 
         xcube_hub_namespace = os.getenv("K8S_NAMESPACE", "xcube-gen-dev")
         user_namespaces.create_if_not_exists(user_namespace=xcube_hub_namespace)
@@ -218,7 +235,7 @@ def delete_all(user_id: str) -> Union[AnyDict, Error]:
         raise api.ApiError(400, str(e))
 
 
-def info(user_id: str, body: JsonObject, token: Optional[str] = None) -> JsonObject:
+def info(user_id: str, email: str, body: JsonObject, token: Optional[str] = None) -> JsonObject:
     data_pools_cfg_file = os.getenv("XCUBE_GEN_DATA_POOLS_PATH", None)
     if data_pools_cfg_file is None:
         raise api.ApiError(400, "XCUBE_GEN_DATA_POOLS_PATH is not configured.")
@@ -229,7 +246,7 @@ def info(user_id: str, body: JsonObject, token: Optional[str] = None) -> JsonObj
     except FileNotFoundError as e:
         raise api.ApiError(400, f"Datapools file {data_pools_cfg_file} not found. " + str(e))
 
-    job = create(user_id=user_id, cfg=body, info_only=True, token=token)
+    job = create(user_id=user_id, email=email, cfg=body, info_only=True, token=token)
     apps_v1_api = client.BatchV1Api()
     poller.poll_job_status(apps_v1_api.read_namespaced_job_status, namespace="xcube-gen-stage",
                            name=job['cubegen_id'])
@@ -254,4 +271,15 @@ def info(user_id: str, body: JsonObject, token: Optional[str] = None) -> JsonObj
     except KeyError:
         raise api.ApiError(400, f'unsupported "input_config/datastore_id" entry: "{store_id}"')
 
-    return costs.get_size_and_cost(processing_request=processing_request, datastore=data_store)
+    available = punits.get_punits(user_id=email)
+
+    cost_est = costs.get_size_and_cost(processing_request=processing_request, datastore=data_store)
+    required = cost_est['punits']['total_count']
+
+    limit = os.getenv("XCUBE_HUB_PROCESS_LIMIT", 1000)
+
+    return dict(
+        dataset_descriptor=cost_est['dataset_descriptor'],
+        size_estimation=cost_est['size_estimation'],
+        cost_estimation=dict(required=required, available=available['count'], limit=limit)
+    )